// multisto/src/Schema.ts
export interface Documentfields {
  ID: number;
  title?: string | null;
  fieldtype?: number | null;
  stoid?: number | null;
  documentid?: number | null;
  fieldid?: string | null;
  fieldhelpertext?: string | null;
  externalFileDataLabel?: string | null;
}
// ...
export interface Documents {
  ID: number;
  title?: string | null;
  contents?: string | null;
  stoid?: number | null;
  directoryid?: number | null;
  /**  Defaults to: 0. */
  isactiveforinvestors: number;
  /**  Defaults to: 0. */
  filetype?: number | null;
  /**  Defaults to: 0. */
  offerID: number;
  isactiveforinvestorsType?: number | null;
  isactiveforinvestorsNames?: string | null;
  /**  Defaults to: ["ALL"]. */
  countriesWhitelist?: string | null;
  docusignDocumentID?: string | null;
  /**  Defaults to: ["ALL"]. */
  investorTypesWhitelist?: string | null;
  helloSignDocumentID?: string | null;
}
// ...
export interface Documentuser {
  ID: number;
  contents?: string | null;
  investorID?: number | null;
  stoid?: number | null;
  directoryid?: number | null;
  documentid: number;
  DocumentStatus: number;
  fieldValuesJson?: string | null;
  documentofferinvestorid: number;
  signaturefilepath?: string | null;
  signaturedate?: Date | null;
  signatureFileID?: number | null;
  sharePurchaseID?: number | null;
}
// ...
export interface Investor {
  ID: number;
  FirstName?: string | null;
  LastName?: string | null;
  Address?: string | null;
  country?: string | null;
  phone?: string | null;
  cell?: string | null;
  zip?: string | null;
  town?: string | null;
  state?: string | null;
  Password: string;
  email: string;
  PassportNumber?: string | null;
  NationalID?: string | null;
  /**  Defaults to: 0. */
  receiveEmails?: number | null;
  kinname?: string | null;
  kinphone?: string | null;
  kinemail?: string | null;
  /**  Defaults to: 0. */
  investorType: number;
  CompanyName?: string | null;
  TitleWithinCompany?: string | null;
  /**  Defaults to: 0. */
  PowerToBindCompany?: number | null;
  DOB?: Date | null;
  /**  Defaults to: 0. */
  twofactorenable?: number | null;
  /**  Defaults to: 0. */
  investorBulkUploadStatus?: number | null;
  /**  Defaults to: en. */
  language?: string | null;
  MiddleName?: string | null;
  SocialSecurity?: string | null;
  MailingAddress?: string | null;
  FaxNumber?: string | null;
  /**  Defaults to: 0. */
  MaritalStatus?: number | null;
  Occupation?: string | null;
  EmployerName?: string | null;
  EmployerAddress?: string | null;
  /**  Defaults to: 0. */
  RetirementAccount?: number | null;
  /**  Defaults to: 0. */
  TrustOrBusinessEntity?: number | null;
  DateIncorporation?: Date | null;
  TaxIDNo?: string | null;
  GovtIDNo?: string | null;
  /**  Defaults to: 0. */
  GovtIDNoIsTaxNo?: number | null;
  NameSecondaryContact?: string | null;
  PhoneSecondaryContact?: string | null;
  EmailSecondaryContact?: string | null;
  NamePrimaryContact?: string | null;
  PhonePrimaryContact?: string | null;
  EmailPrimaryContact?: string | null;
  PrincipalCountryOfBusiness?: string | null;
  countryOfCitizenship?: string | null;
  /**  Defaults to: 0. */
  referByInvestorID?: number | null;
  kyc?: string | null;
  /**  Defaults to: 1. */
  dividendPeriod?: number | null;
  DOBCountry?: string | null;
  taxResidencyCountry?: string | null;
  /**  Defaults to: 0. */
  affiliateStatus: number;
  /**  Defaults to: 0. */
  PoliticallyExposedPerson: number;
  PoliticallyExposedPersonPost?: string | null;
  mercuryID?: string | null;
  uploadDate?: Date | null;
  twoFactorCode?: number | null;
  brokerID?: string | null;
  driversLicenseID?: string | null;
  /**  Defaults to: CURRENT_TIMESTAMP. */
  investmentLimitUpdateDate: Date;
  /**  Defaults to: 0. */
  allowedTotalInvestment: number;
  /**  Defaults to: 0. */
  yearlyTotalInvestment: number;
}
// ...
export interface Stos {
  ID: number;
  title: string;
  details: string;
  /**  Defaults to: 1. */
  isActive?: number | null;
  logo: string;
  ethereumContractAddress: string;
  ethereumWhitelistAddress: string;
  disclamer?: string | null;
  stolink?: string | null;
  stolinkfull?: string | null;
  /**  Defaults to: 0. */
  stoType?: number | null;
  stoinvestortypes?: string | null;
  emailFooter?: string | null;
  steps?: string | null;
  registrationtext?: string | null;
  SMTP_Host?: string | null;
  SMTP_Port?: string | null;
  SMTP_User?: string | null;
  SMTP_Password?: string | null;
  SMTP_FromAddress?: string | null;
  website?: string | null;
  stoinvestortypesnotonshareregister?: string | null;
  /**  Defaults to: 0. */
  companytype: number;
  settings?: string | null;
  registrationsuccesstext?: string | null;
  tellafriendtext?: string | null;
  inviteFriendEmailText?: string | null;
  PropertyFullDetails?: string | null;
  /**  Defaults to: 2030-01-01. */
  exchangeOpenDate?: Date | null;
  propertypicture?: string | null;
  docusign_sto_contract?: string | null;
  docusign_sto_purchase?: string | null;
  /**  Defaults to: 0. */
  externalSystemID?: number | null;
  projectAddress?: string | null;
  LegalDetails?: string | null;
  affiliatePlanId?: number | null;
  affiliateShareTypeId?: number | null;
  /**  Defaults to: 0. */
  isICOShareTypeCompany?: number | null;
  EmailTxtInvestorBulkUpload?: string | null;
  /**  Defaults to: 1. */
  isBuyButtonEnabled?: number | null;
  isBimountEnabled?: number | null;
  /**  Defaults to: 0. */
  projectCost?: number | null;
  VerifyInvestorComHostToken?: string | null;
  helloSignClientID?: string | null;
  SMTP_FromName?: string | null;
  /**  Defaults to: CURRENT_TIMESTAMP. */
  createdAt: Date;
  /**  Defaults to: 0. */
  defaultBlockchain?: number | null;
}